car stats in car files
