+++ nfm-world/data/models/cars/2000tornados.rad
@@ -1,3 +1,28 @@
+swits(50, 185, 282)
+acelf(11.0, 5.0, 3.0)
+handb(7)
+airs(1.0)
+airc(70)
+turn(6)
+grip(20.0)
+bounce(1.2)
+simag(0.9)
+moment(1.3)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(0)
+revlift(0)
+powerloss(2500000)
+flipy(-50)
+msquash(7)
+clrad(3300)
+dammult(0.75)
+maxmag(7600)
+dishandle(0.65)
+outdam(0.68)
+enginsignature(0)
+
 idiv(205)
 shadow()
 
@@ -647,7 +672,7 @@ p(-31,11,-58)
 </p>
 
 
-//rims([color],size,depth)
+//rims((color),size,depth)
 
 rims(140,140,140,20,20)
