+++ nfm-world/data/models/cars/audir8.rad
@@ -1,3 +1,27 @@
+swits(90, 200, 305)
+acelf(13.0, 7.0, 4.5)
+handb(8)
+airs(1.0)
+airc(50)
+turn(6)
+grip(30.0)
+bounce(1.05)
+simag(0.9)
+moment(1.26)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(10)
+revlift(0)
+powerloss(3500000)
+flipy(-28)
+msquash(3)
+clrad(10000)
+dammult(0.61)
+maxmag(13000)
+dishandle(0.87)
+outdam(0.77)
+
 idiv(160)
 shadow()
