+++ nfm-world/data/models/cars/canyenaro.rad
@@ -1,3 +1,27 @@
+swits(60, 180, 275)
+acelf(10.0, 5.0, 3.5)
+handb(7)
+airs(0.95)
+airc(40)
+turn(5)
+grip(18.0)
+bounce(1.3)
+simag(1.05)
+moment(1.4)
+comprad(0.8)
+push(3)
+revpush(2)
+lift(0)
+revlift(15)
+powerloss(3500000)
+flipy(-92)
+msquash(7)
+clrad(4700)
+dammult(0.45)
+maxmag(7200)
+dishandle(0.55)
+outdam(0.8)
+
 div(19)
 shadow()
