+++ nfm-world/data/models/cars/drifter.rad
@@ -1,3 +1,27 @@
+swits(90, 210, 295)
+acelf(12.0, 7.0, 4.0)
+handb(5)
+airs(1.0)
+airc(40)
+turn(9)
+grip(19.0)
+bounce(1.2)
+simag(1.0)
+moment(1.35)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(20)
+revlift(0)
+powerloss(2750000)
+flipy(-77)
+msquash(3)
+clrad(10000)
+dammult(0.58)
+maxmag(18000)
+dishandle(0.45)
+outdam(0.72)
+
 idiv(157)
 shadow()
