+++ nfm-world/data/models/cars/drmonster.rad
@@ -1,3 +1,27 @@
+swits(70, 210, 290)
+acelf(12.0, 6.0, 3.5)
+handb(7)
+airs(1.0)
+airc(60)
+turn(6)
+grip(27.0)
+bounce(1.15)
+simag(1.15)
+moment(2.0)
+comprad(0.8)
+push(2)
+revpush(1)
+lift(0)
+revlift(32)
+powerloss(5500000)
+flipy(-127)
+msquash(8)
+clrad(5000)
+dammult(0.52)
+maxmag(18000)
+dishandle(0.95)
+outdam(1.0)
+
 idiv(182)
 shadow()
