+++ nfm-world/data/models/cars/king.rad
@@ -1,3 +1,27 @@
+swits(50, 160, 270)
+acelf(9.0, 5.0, 3.0)
+handb(10)
+airs(0.8)
+airc(10)
+turn(4)
+grip(25.0)
+bounce(0.8)
+simag(1.1)
+moment(2.0)
+comprad(1.5)
+push(4)
+revpush(1)
+lift(0)
+revlift(0)
+powerloss(4500000)
+flipy(-85)
+msquash(10)
+clrad(7000)
+dammult(0.45)
+maxmag(10700)
+dishandle(0.4)
+outdam(0.95)
+
 div(19)
 shadow()
