+++ nfm-world/data/models/cars/koolkat.rad
@@ -1,3 +1,27 @@
+swits(70, 206, 291)
+acelf(11.0, 5.0, 3.0)
+handb(10)
+airs(0.8)
+airc(90)
+turn(5)
+grip(20.0)
+bounce(1.1)
+simag(0.9)
+moment(1.48)
+comprad(0.5)
+push(4)
+revpush(1)
+lift(0)
+revlift(0)
+powerloss(3200000)
+flipy(-60)
+msquash(4)
+clrad(5000)
+dammult(0.6)
+maxmag(17000)
+dishandle(0.72)
+outdam(0.76)
+
 idiv(180)
 shadow()
