+++ nfm-world/data/models/cars/leadoxide.rad
@@ -1,3 +1,27 @@
+swits(60, 170, 289)
+acelf(7.0, 9.0, 4.0)
+handb(9)
+airs(0.9)
+airc(40)
+turn(5)
+grip(25.0)
+bounce(1.15)
+simag(1.05)
+moment(1.43)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(0)
+revlift(0)
+powerloss(3200000)
+flipy(-54)
+msquash(6)
+clrad(3500)
+dammult(0.72)
+maxmag(17200)
+dishandle(0.6)
+outdam(0.82)
+
 div(17)
 shadow()
