+++ nfm-world/data/models/cars/lescrab.rad
@@ -1,3 +1,27 @@
+swits(76, 195, 298)
+acelf(11.0, 6.0, 3.5)
+handb(15)
+airs(1.0)
+airc(40)
+turn(7)
+grip(22.0)
+bounce(1.15)
+simag(0.9)
+moment(1.2)
+comprad(0.5)
+push(3)
+revpush(2)
+lift(20)
+revlift(0)
+powerloss(2500000)
+flipy(-44)
+msquash(2)
+clrad(3000)
+dammult(0.8)
+maxmag(6000)
+dishandle(0.77)
+outdam(0.5)
+
 div(17)
 shadow()
