+++ nfm-world/data/models/cars/masheen.rad
@@ -1,3 +1,27 @@
+swits(50, 130, 210)
+acelf(7.5, 3.5, 3.0)
+handb(12)
+airs(0.3)
+airc(0)
+turn(5)
+grip(27.0)
+bounce(0.8)
+simag(1.3)
+moment(3.0)
+comprad(0.8)
+push(2)
+revpush(2)
+lift(0)
+revlift(0)
+powerloss(16700000)
+flipy(-100)
+msquash(20)
+clrad(15000)
+dammult(0.25)
+maxmag(45000)
+dishandle(0.42)
+outdam(1.0)
+
 idiv(295)
 iwid(117)
 shadow()
