+++ nfm-world/data/models/cars/maxrevenge.rad
@@ -1,3 +1,27 @@
+swits(70, 202, 293)
+acelf(12.0, 6.0, 3.0)
+handb(8)
+airs(1.0)
+airc(50)
+turn(7)
+grip(20.0)
+bounce(1.2)
+simag(0.9)
+moment(1.38)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(30)
+revlift(0)
+powerloss(2500000)
+flipy(-57)
+msquash(4)
+clrad(4500)
+dammult(0.7)
+maxmag(15000)
+dishandle(0.9)
+outdam(0.76)
+
 idiv(147)
 shadow()
