+++ nfm-world/data/models/cars/mustang.rad
@@ -1,3 +1,27 @@
+swits(70, 200, 295)
+acelf(11.5, 6.5, 3.5)
+handb(8)
+airs(1.15)
+airc(75)
+turn(7)
+grip(22.5)
+bounce(1.15)
+simag(0.9)
+moment(1.42)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(0)
+revlift(0)
+powerloss(2750000)
+flipy(-82)
+msquash(4)
+clrad(4000)
+dammult(0.67)
+maxmag(19000)
+dishandle(0.95)
+outdam(0.79)
+
 idiv(139)
 shadow()
