+++ nfm-world/data/models/cars/nimi.rad
@@ -1,3 +1,27 @@
+swits(70, 170, 275)
+acelf(10.0, 5.0, 3.5)
+handb(12)
+airs(2.2)
+airc(30)
+turn(8)
+grip(19.0)
+bounce(1.3)
+simag(0.85)
+moment(1.1)
+comprad(0.4)
+push(2)
+revpush(2)
+lift(0)
+revlift(0)
+powerloss(4000000)
+flipy(-60)
+msquash(8)
+clrad(2000)
+dammult(0.42)
+maxmag(6000)
+dishandle(0.62)
+outdam(0.42)
+
 div(19)
 shadow()
