+++ nfm-world/data/models/cars/policecops.rad
@@ -1,3 +1,27 @@
+swits(90, 190, 276)
+acelf(12.0, 7.0, 3.5)
+handb(7)
+airs(0.9)
+airc(50)
+turn(7)
+grip(24.0)
+bounce(1.1)
+simag(1.05)
+moment(1.7)
+comprad(0.8)
+push(2)
+revpush(1)
+lift(0)
+revlift(0)
+powerloss(5500000)
+flipy(-57)
+msquash(8)
+clrad(15000)
+dammult(0.41)
+maxmag(11000)
+dishandle(0.8)
+outdam(0.62)
+
 idiv(170)
 shadow()
