+++ nfm-world/data/models/cars/radicalone.rad
@@ -1,3 +1,27 @@
+swits(80, 200, 300)
+acelf(11.0, 7.5, 4.0)
+handb(7)
+airs(1.3)
+airc(100)
+turn(7)
+grip(25.0)
+bounce(1.1)
+simag(0.9)
+moment(1.5)
+comprad(0.5)
+push(2)
+revpush(2)
+lift(30)
+revlift(0)
+powerloss(3000000)
+flipy(-63)
+msquash(3)
+clrad(5500)
+dammult(0.38)
+maxmag(5800)
+dishandle(1.0)
+outdam(0.85)
+
 div(16)
 shadow()
