+++ nfm-world/mad/ContO.cs
@@ -78,15 +78,15 @@ public class ContO
 
     public static implicit operator ContO(Mesh mesh) => new ContO(mesh);
 
-    public void DamageX(Stat stat, int wheelnum, float amount)
+    public void DamageX(CarStats stat, int wheelnum, float amount)
     {
         MeshDamage.DamageX(stat, _mesh, wheelnum, amount);
     }
-    public void DamageY(Stat stat, int wheelnum, float amount, bool mtouch, int nbsq, int squash)
+    public void DamageY(CarStats stat, int wheelnum, float amount, bool mtouch, int nbsq, int squash)
     {
         MeshDamage.DamageY(stat, _mesh, wheelnum, amount, mtouch, ref nbsq, ref squash);
     }
-    public void DamageZ(Stat stat, int wheelnum, float amount)
+    public void DamageZ(CarStats stat, int wheelnum, float amount)
     {
         MeshDamage.DamageZ(stat, _mesh, wheelnum, amount);
     }
